+++ r/R/dplyr-funcs-datetime.R
@@ -350,6 +350,24 @@ register_bindings_duration <- function() {
     delta <- build_expr("floor", seconds * fraction)
     delta <- delta$cast(int64())
     start + delta$cast(duration("s"))
+  register_binding("dseconds", function(x = 1) {
+    dseconds_int64 <- Expression$create("cast", x, options = cast_options(to_type = int64()))
+    build_expr("cast", dseconds_int64, options = list(to_type = duration(unit = "s")))
+  })
+  register_binding("dmilliseconds", function(x = 1) {
+    dmilliseconds_int64 <- Expression$create("cast", x, options = cast_options(to_type = int64()))
+    build_expr("cast", dmilliseconds_int64, options = list(to_type = duration(unit = "ms")))
+  })
+  register_binding("dmicroseconds", function(x = 1) {
+    dmicroseconds_int64 <- Expression$create("cast", x, options = cast_options(to_type = int64()))
+    build_expr("cast", dmicroseconds_int64, options = list(to_type = duration(unit = "us")))
+  })
+  register_binding("dnanoseconds", function(x = 1) {
+    dnanoseconds_int64 <- Expression$create("cast", x, options = cast_options(to_type = int64()))
+    build_expr("cast", dnanoseconds_int64, options = list(to_type = duration(unit = "ns")))
+  })
+  register_binding("dpicoseconds", function(x = 1) {
+    arrow_not_supported("Duration in picoseconds.")
   })
 }